join_m2m method replacement
